hcl/parser: don't print "illegal"
We shouldn't output any text during parsing, that's just strange!

token.ILLEGAL happens when the scanner has an error. This error is
recorded in `scerr` during parsing which is eventually returned out.
Therefore, we handle the error properly. We can simply return a backup
error here in case we don't (scanner errors are checked first).