cmd/zclfmt: zcl native syntax pretty-printer
This applies the simple native syntax reformatting function to one or
more files. It does not support JSON or any other syntax.

Calling applications might provide their own versions of this that e.g.
can format an entire directory by matching on filename patterns, but
this serves as an example and a utility for single files.