printer: Add support for external filters during Format
This PR introduces the ability to pass in external filters to alter the
formatting process in a way specific to the tool that HCL is implemented
for. An example would be pre-populating fields in Terraform during
terraform fmt, such as the description fields for variables and outputs.

printer.Filter is specifically an interface - the downstream tool
supplies the Filter method, which carries out the necessary behaviour.

Also in this commit are some updates to docs for Run and Options in
fmtcmd.