hclsyntax: allow missing newline at EOF
Due to some earlier limitations of the parser we required each attribute
and block to end with a newline, even if it appeared at the end of a
file. In effect, this required all files to end with a newline character.

This is no longer required and so we'll tolerate that missing newline for
pragmatic reasons.